fix(socialaccount): make to_str return unique identifier
The view allauth.socialaccount.views.connections displays a list of
social accounts to the user. Instead of using the display name of a
social account, which may not be unique, use the unique username or
email address instead. That way, accounts with identical display names
can be distinguished from each other.

Fixes #3906